requirements.txt
# Real vs AI Face Detection - Dependencies

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
Pillow>=9.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Model Architectures (optional - for additional models)
timm>=0.9.0  # PyTorch Image Models (EfficientNet, etc.)

# Experiment Tracking (optional)
# tensorboard>=2.13.0
# wandb>=0.15.0

# Jupyter Support (optional)
# jupyter>=1.0.0
# ipywidgets>=8.0.0
